REnamed Gtests and removed old test files
