+++ tests/cpp/test1.cpp
@@ -1,32 +1,26 @@
-/**
- * Nullity test of Divergence operator
- */
-
 #include "mole.h"
-#include <iostream>
+#include <gtest/gtest.h>
 
 void run_nullity_test(int k, Real tol) {
-  int m = 2 * k + 1;
-  Real dx = 1;
+    int m = 2 * k + 1;
+    Real dx = 1;
 
-  Divergence D(k, m, dx);
-  vec field(m + 1, fill::ones);
+    Divergence D(k, m, dx);
+    vec field(m + 1, fill::ones);
 
-  vec sol = D * field;
+    vec sol = D * field;
 
-  if (norm(sol) > tol) {
-    cout << "\033[1;31mTest FAILED!\033[0m\n";
-    exit(1);
-  }
+    EXPECT_NEAR(norm(sol), 0, tol);
 }
 
-int main() {
-  Real tol = 1e-10;
+TEST(DivergenceTests, Nullity) {
+    Real tol = 1e-10;
+    for (int k : {2, 4, 6}) {
+        run_nullity_test(k, tol);
+    }
+}
 
-  for (int k : {2, 4, 6})
-    run_nullity_test(k, tol);
-
-  cout << "\033[1;32mTest PASSED!\033[0m\n";
-
-  return 0;
+int main(int argc, char **argv) {
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
 }
